bug: fix FN980 modem detection
FN980 modem dection compares the value of _profile variable to 0 to
check whether FN980_QUIRK has to be activated, however at the moment of
the check _profile could be undefined, so the check fails.

Since the FN980_QUIRK override the value of _profile anyway, there is no
need to check its current value, so the check can be completely removed,
fixing the above error.

Change-Id: Id8a73df12363c48b1c16f028270c6c6d231808f9