bug: delete wds profile prevent a second connection
Currently, tqmi-connect scripts delete the newly created profile at the
end of the script, also in case of errors, however this will also
prevent a second connection made with the same script to working
properly (probably inner levels are fooled by using the same profile
id).

To avoid this, with this change the command to delete the profile is
only printed out at the end of the script, and the user is encouraged to
clean up the profile manually.

Change-Id: I8d4c2bdc06b2df30cbdf05a196e137ef3ad123e3