help message: add NOTE about the type of iface needed for the --release flag
currently, it is not clear enough that the --connect and --release flags
require a different type of --iface parameter.

--connect wants the main interface name, which is usually wwan0, and in
fact --iface defaults to wwan0.

--release wants qmimux interface name (e.g. qmimux0), because it is used
to find the session configuration file, which is named after the qmimux
interface created (e.g. ip-session-qmimux0).

This commit adds an "IMPORTANT NOTE" message just above the --release
example that shows the correct usage.

Change-Id: I7859366f5bc7838335db2265bcd10e32676bc0e1