fix: call wds-delete-profile only if there is a profile set
currently, wds-delete-profile is called even if the script did not set
any profile (e.g. with LM960 modem).

This change avoid calling wds-delete-profile if a profile was not set,
checking both $_profile, which comes from the command line, and
$PROFILE, which comes from the configuration file instead, since the
profile could be deleted in a normal --release (so we need PROFILE from
the configuration file), and in case of errors during both --release and
--connect (so we also need to check $_profile).

Change-Id: I5b6ef27c54c59b7940eb6ddc60df1dc11269e4d9