tqmi-connect: use dl aggregated max size returned by the modem as MTU
When enabling dl data aggregation the URB sent by the host should be
at least as large as the dl aggregated max size returned by the modem.

Due to qmi_wwan and usbnet behavior, this result can be achieved by
setting the MTU of the master network interface from which the virtual
ones are derived.

Instead of using the default mtu or the one provided as argument from
the command line use the one returned by the modem because it's the
correct one.

Keep also the argument, since it's useful as an hint for the modem and
to simplify testing with different URB sizes.

Change-Id: I00adf7a97f6fe40f1e3c13879a6f46b165604277